+++ R/foreign.R
@@ -1,42 +1,43 @@
-sp2map <- function(x) {
-  g <- bld2(x, normalize_verts = FALSE)
-  
-  v <- g$o %>% inner_join(g$b) %>% inner_join(g$bXv) %>% inner_join(g$v) %>% 
- # v <- x$v %>% inner_join(x$bXv) %>% inner_join(x$b) %>% inner_join(x$o) %>% 
-    mutate(obj = paste(.ob0, .br0, sep = "_")) %>% group_by(obj) %>% 
-    do(rbind(., NA_real_))
-  v <- v[-nrow(v), ]
-  m <- list(x = v$x, y = v$y, range = c(range(v$x, na.rm = TRUE), range(v$y, na.rm = TRUE)), 
-            names = unique(v$objs))
-
-  class(m) <- "map"
-  m
-}
-
-map2gris <- function(x, orphans = TRUE) {
-  na <- is.na(x$x)
-  segs <- c(0, cumsum(abs(diff(is.na(x$x)))))
-  nms <- x$names
-  
-  y <- x$y
-  x <- x$x
-
-    ##clist <- split(data.frame(x$x, x$y)[!na, ], segs[!na])
-  v <- data_frame(x = x, y = y, .br0 = segs) %>% filter(!is.na(x)) %>% mutate(.vx0 = seq(n()))
-  bXv <- v %>% select(.br0, .vx0)
-  v <- v %>% select(x, y, .vx0)
-  o <- normalizeVerts2(v, bXv, c("x", "y"))
-  # #may need keep_all here
-  o$b <- bXv %>% dplyr::distinct(.br0) %>% select(.br0) 
-  dif <- length(nms) - nrow(o$b)
-  if (dif < 0) {
-    nms <- c(nms, paste("orphan", seq(-dif)))
-  }
-  o$o <- data_frame(names = nms, .ob0 = seq(length(nms)))
-  o$b$.ob0 <- o$o$.ob0
-  class(o) <- c("gris", "list")
-  o
-}
+# # these won't work yet (changes to names x_/y_ and auto normalize by bld3) 
+#sp2map <- function(x) {
+#   g <- bld3(x, normalize_verts = FALSE)
+#   
+#   v <- g$o %>% inner_join(g$b) %>% inner_join(g$bXv) %>% inner_join(g$v) %>% 
+#  # v <- x$v %>% inner_join(x$bXv) %>% inner_join(x$b) %>% inner_join(x$o) %>% 
+#     mutate(obj = paste(object_, branch_, sep = "_")) %>% group_by(obj) %>% 
+#     do(rbind(., NA_real_))
+#   v <- v[-nrow(v), ]
+#   m <- list(x = v$x, y = v$y, range = c(range(v$x, na.rm = TRUE), range(v$y, na.rm = TRUE)), 
+#             names = unique(v$objs))
+# 
+#   class(m) <- "map"
+#   m
+# }
+# 
+# map2gris <- function(x, orphans = TRUE) {
+#   na <- is.na(x$x)
+#   segs <- c(0, cumsum(abs(diff(is.na(x$x)))))
+#   nms <- x$names
+#   
+#   y <- x$y
+#   x <- x$x
+# 
+#     ##clist <- split(data.frame(x$x, x$y)[!na, ], segs[!na])
+#   v <- data_frame(x = x, y = y, branch_ = segs) %>% filter(!is.na(x)) %>% mutate(vertex_ = seq(n()))
+#   bXv <- v %>% select(branch_, vertex_)
+#   v <- v %>% select(x, y, vertex_)
+#   o <- normalizeVerts2(v, bXv, c("x", "y"))
+#   # #may need keep_all here
+#   o$b <- bXv %>% dplyr::distinct(branch_) %>% select(branch_) 
+#   dif <- length(nms) - nrow(o$b)
+#   if (dif < 0) {
+#     nms <- c(nms, paste("orphan", seq(-dif)))
+#   }
+#   o$o <- data_frame(names = nms, object_ = seq(length(nms)))
+#   o$b$object_ <- o$o$object_
+#   class(o) <- c("gris", "list")
+#   o
+# }
 
 
 #' @importFrom sp Polygon Line Polygons SpatialPolygons Lines SpatialLines
@@ -55,13 +56,13 @@ vertsToPoint <- function(v) {
 grisToSPbase <- function(x, type = "pp") {
   conv <- switch(type, pp = vertsToPoly, l = vertsToLine, p = vertsToPoint)
     branches <- x$v %>% 
-    inner_join(x$bXv, by = c(".vx0" = ".vx0")) %>% 
-    inner_join(x$b, by = c(".br0" = ".br0"))
-  lapply(split(branches, branches$.br0), conv)
+    inner_join(x$bXv, by = c("vertex_" = "vertex_")) %>% 
+    inner_join(x$b, by = c("branch_" = "branch_"))
+  lapply(split(branches, branches$branch_), conv)
 }
 
 grisToSpatialPolygons <- function(x) 
-  SpatialPolygons(lapply(x$o$.ob0, function(obid) Polygons(grisToSPbase(x[which(x$o$.ob0 == obid), ]), obid)))
+  SpatialPolygons(lapply(x$o$object_, function(obid) Polygons(grisToSPbase(x[which(x$o$object_ == obid), ]), obid)))
 
 #' Return a list of Polygon, Line or matrix from gris
 #'
@@ -85,10 +86,10 @@ as.SpatialPolygonsDataFrame <- function(x, ...) {
 #' @importFrom sp Polygons SpatialPolygons SpatialPolygonsDataFrame
 as.SpatialPolygonsDataFrame.gris <- function(x, ...) {
   geom <- grisToSpatialPolygons(x)
-  dat <- x$o
-  rownames(dat) <- dat$.ob0
-  dat$.ob0 <- NULL
-  SpatialPolygonsDataFrame(geom, as.data.frame(dat))
+  dat <- as.data.frame(x$o)
+  rownames(dat) <- dat$object_
+  dat$object_ <- NULL
+  SpatialPolygonsDataFrame(geom, dat)
 }